partial mitigation of https://www.lunasec.io/docs/blog/log4j-zero-day/
First I thought that:

| This looks more like someone at Minecraft did the equivalent of
| printf(user_controlled_string) instead of
| printf("%s", user_controlled_string) though, from the description.

But it’s actually that THE LOGGING LIBRARY ITSELF does that (interpret
user data in the to-be-logged message), and in 2016 someone already said
https://issues.apache.org/jira/browse/LOG4J2-905 that’s a bad idea, and
all it got was a backwards-incompatible mitigation (use %msg{nolookups}
which writes the extra string “{nolookups}” (literally) after your log
message in older releases).

This is so messed up, conceptually.

I wonder when it was introduced and whether parts apply to log4j version 1…